moved dta loading to main
